sauvegarde + chargement local

+++ app/src/main/java/com/example/chief/data/model/Ingredient.kt
@@ -2,6 +2,6 @@ package com.example.chief.data.model
 
 data class Ingredient(
     val nom: String,
-    val quantite: Int,
+    val quantite: Float,
     val unite: String
 )
